requirements.txt
svglib==0.9.4
gevent==21.1.0
flask==2.2.3
reportlab==3.6.1
jinja2==3.1.2
pillow==9.4.0
